chore(ci)!: retire the skill-overlap ratchet (#511)
Delete .github/workflows/skill-overlap.yml, tools/skill-overlap/, both
.github/skill-overlap-*.yml contracts, and the workflow's contract test.
Drop the test-skill-overlap leg from `just check` and the Rust target
ignore.

Evidence for retiring rather than salvaging:

- It could never enforce. The main-protection ruleset carries deletion,
  non_fast_forward, pull_request, and copilot_code_review — no
  required_status_checks — and branches/main/protection is 404 (#340).
  There is no merge queue and no merge_group trigger, so trunk had no
  invariant either (#341). Both gaps need an owner-level merge-policy
  change, not a code change.
- It was never calibrated. Calibration and baseline were both
  `status: draft` with every detector digest and every labeled sample
  literally `review-required`, so ADR-004's blocking `check` mode was
  unreachable by construction.
- It was expensive and unread. Every PR touching skills/**/*.md paid a
  full Rust build, a FastEmbed model fetch, and a golden-vector parity
  check under a 30-minute timeout to produce an advisory report with no
  consumer; every historical run is a green `calibrate`.
- It was a maintenance liability. 4,783 lines of Rust plus a Cargo.lock
  that dependabot.yml does not cover, pinning FastEmbed 5.17.3, and
  `just test-skill-overlap` was the only leg of `just check` requiring a
  Rust toolchain. The gate is now Python, JS, and bash only.

The removal commit is the restore point: `git revert` brings the
analyzer back whole, which is more recoverable than the parked
dispatch-only workflow it replaces.

tools/skill-overlap/fixtures/relative-md-refs.json existed to hold the
Python/Rust parity cases for relative_md_refs. With the Rust side gone
the cases move inline into tests/python/test_ref_extraction.py (renamed
from test_ref_extraction_fixture.py) as parametrized cases, so the
reference-resolution gates that depend on that extractor keep their
coverage — including the pinned trailing-newline quirk.

Wiki: ADR skill-overlap-ratchet-005 records the decision and its
alternatives; 001-004 are marked superseded rather than deleted, and
tooling.md drops the retired workflow and gate leg.
