--- .github/workflows/skill-overlap.yml
@@ -1,162 +1,0 @@
-name: skill-overlap
-
-# Disabled for now: automatic schedule and pull_request triggers are parked
-# while we decide whether to salvage the ratchet. Manual runs still work via
-# workflow_dispatch. See https://github.com/paulnsorensen/easy-cheese/issues/511.
-on:
-  workflow_dispatch:
-    inputs:
-      mode:
-        description: Analysis mode
-        required: true
-        default: calibrate
-        type: choice
-        options: [calibrate, report, check]
-
-permissions:
-  contents: read
-
-concurrency:
-  group: ${{ github.workflow }}-${{ github.ref }}
-  cancel-in-progress: true
-
-jobs:
-  overlap:
-    name: skill overlap ratchet
-    runs-on: ubuntu-latest
-    timeout-minutes: 30
-    env:
-      CARGO_TERM_COLOR: always
-    steps:
-      - uses: actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0 # v7.0.0
-        with:
-          persist-credentials: false
-      - name: Set MODEL_DIR
-        run: echo "MODEL_DIR=$RUNNER_TEMP/skill-overlap-model" >> "$GITHUB_ENV"
-      - uses: dtolnay/rust-toolchain@4cda84d5c5c54efe2404f9d843567869ab1699d4 # stable
-      - name: Restore cargo + model cache
-        id: cache
-        uses: actions/cache/restore@0057852bfaa89a56745cba8c7296529d2fc39830 # v4.3.0
-        with:
-          path: |
-            ~/.cargo/registry
-            ~/.cargo/git
-            tools/skill-overlap/target
-            ${{ env.MODEL_DIR }}
-          key: skill-overlap-${{ runner.os }}-${{ hashFiles('tools/skill-overlap/Cargo.lock', 'tools/skill-overlap/model.lock.json') }}
-          restore-keys: |
-            skill-overlap-${{ runner.os }}-
-      - name: Select advisory or requested mode
-        id: mode
-        shell: bash
-        run: |
-          set -euo pipefail
-          requested="${{ inputs.mode || '' }}"
-          if [[ -n "$requested" ]]; then
-            mode="$requested"
-          else
-            calibration_status=$(awk '$1 == "status:" { print $2; exit }' .github/skill-overlap-calibration.yml)
-            baseline_status=$(awk '$1 == "status:" { print $2; exit }' .github/skill-overlap-baseline.yml)
-            case "$calibration_status" in
-              draft)
-                if [[ "$baseline_status" != draft ]]; then
-                  echo "baseline must remain draft until calibration is reviewed" >&2
-                  exit 1
-                fi
-                mode=calibrate
-                ;;
-              reviewed)
-                if ! cargo run --quiet --manifest-path tools/skill-overlap/Cargo.toml -- calibration validate \
-                  --calibration .github/skill-overlap-calibration.yml \
-                  --model-lock tools/skill-overlap/model.lock.json; then
-                  echo "reviewed calibration failed validation" >&2
-                  exit 1
-                fi
-                case "$baseline_status" in
-                  draft)
-                    mode=report
-                    ;;
-                  reviewed)
-                    if ! cargo run --quiet --manifest-path tools/skill-overlap/Cargo.toml -- baseline validate \
-                      --baseline .github/skill-overlap-baseline.yml \
-                      --calibration .github/skill-overlap-calibration.yml \
-                      --model-lock tools/skill-overlap/model.lock.json; then
-                      echo "reviewed baseline failed validation" >&2
-                      exit 1
-                    fi
-                    if [[ "${{ github.event_name }}" == "pull_request" ]]; then
-                      mode=check
-                    else
-                      mode=report
-                    fi
-                    ;;
-                  *)
-                    echo "baseline status must be draft or reviewed" >&2
-                    exit 1
-                    ;;
-                esac
-                ;;
-              *)
-                echo "calibration status must be draft or reviewed" >&2
-                exit 1
-                ;;
-            esac
-          fi
-          echo "value=$mode" >> "$GITHUB_OUTPUT"
-      - name: Run model-free analyzer tests
-        run: cargo test --manifest-path tools/skill-overlap/Cargo.toml
-      - name: Fetch and verify locked model
-        run: cargo run --features model --manifest-path tools/skill-overlap/Cargo.toml -- model fetch --model-lock tools/skill-overlap/model.lock.json --model-dir "$MODEL_DIR"
-      - name: Verify local embedding parity
-        run: >-
-          cargo run --features model --manifest-path tools/skill-overlap/Cargo.toml -- verify-parity
-          --model-lock tools/skill-overlap/model.lock.json --model-dir "$MODEL_DIR"
-          --fixture tools/skill-overlap/fixtures/hallouminate-fastembed.json
-      - name: Analyze
-        id: analyze
-        run: |
-          args=(analyze --mode "${{ steps.mode.outputs.value }}" --repo .)
-          args+=(--manifest .claude-plugin/plugin.json)
-          args+=(--model-lock tools/skill-overlap/model.lock.json --model-dir "$MODEL_DIR")
-          args+=(--json-out report.json --markdown-out report.md)
-          if [[ "${{ steps.mode.outputs.value }}" != calibrate ]]; then args+=(--calibration .github/skill-overlap-calibration.yml); fi
-          if [[ "${{ steps.mode.outputs.value }}" == check ]]; then args+=(--baseline .github/skill-overlap-baseline.yml); fi
-          cargo run --features model --manifest-path tools/skill-overlap/Cargo.toml -- "${args[@]}"
-      - name: Job summary
-        if: always()
-        shell: bash
-        run: |
-          {
-            echo "## Skill overlap"
-            echo "- Mode: \`${{ steps.mode.outputs.value }}\`"
-            if [[ -f report.json ]]; then
-              jq -r '"- Report findings: \(.findings | length)\n- Trend groups: \(.trends.groups | length)"' report.json
-            else
-              echo "- Report unavailable; inspect the failed step logs."
-            fi
-            echo "- Full evidence: \`skill-overlap-report\` artifact"
-          } >> "$GITHUB_STEP_SUMMARY"
-          if [[ "${{ steps.analyze.outcome }}" == failure ]]; then
-            echo "::error::skill-overlap check failed; accept a blocker by recording it as intentional or debt with a reason in .github/skill-overlap-baseline.yml, or fix the overlap."
-          fi
-      # Saving is what makes the split worthwhile: it still runs when analysis fails, which is
-      # the run this workflow exists to produce. Skipped on an exact-key hit, where save errors.
-      - name: Save cargo + model cache
-        if: always() && steps.cache.outputs.cache-hit != 'true'
-        uses: actions/cache/save@0057852bfaa89a56745cba8c7296529d2fc39830 # v4.3.0
-        with:
-          path: |
-            ~/.cargo/registry
-            ~/.cargo/git
-            tools/skill-overlap/target
-            ${{ env.MODEL_DIR }}
-          key: skill-overlap-${{ runner.os }}-${{ hashFiles('tools/skill-overlap/Cargo.lock', 'tools/skill-overlap/model.lock.json') }}
-      - uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a # v7.0.1
-        if: always()
-        with:
-          name: skill-overlap-report
-          path: |
-            report.json
-            report.md
-          retention-days: 5
-          if-no-files-found: warn
