+++ .gitignore
@@ -36,9 +36,6 @@ node_modules/
 .idea/
 .vscode/
 
-# Rust build artifacts
-tools/skill-overlap/target/
-
 # Starlight build output and generated docs
 dist/
 .astro/
