+++ .hallouminate/wiki/adr/skill-overlap-ratchet-002.md
@@ -1,5 +1,7 @@
 # ADR: Preserve H2 and H3 identity during semantic chunking
 
+Status: superseded (2026-08-30) by [[skill-overlap-ratchet-005]] — the ratchet is retired and `tools/skill-overlap/` is deleted.
+
 Meaningful H2 and H3 sections remain independent semantic units at any length; only oversized embedding payloads are subdivided, and generated parts retain their original section identity.
 
 ## Decision record
