+++ .hallouminate/wiki/log.md
@@ -62,3 +62,5 @@
 - 2026-08-28 · cd6ee3e392a91bbd · merged · architecture/pyz-bundling-pipeline.md · corrected canonical archive comparison: validate bootstrap members, canonicalize environment and wrapper shebangs, and exclude RECORD only from canonical comparison.
 - 2026-08-28 · cd6ee3e392a91bbd · merged · wiki-conventions.md · adopted the Google developer documentation style guide after repository-specific guidance and recorded its core editorial conventions.
 - 2026-08-28 · 21598ff6fb766b9f · merged · architecture/pyz-bundling-pipeline.md · incorporated PR #501's static `Command` manifests, lazy direct dispatch, command-local argument contract, stdout/stderr channels, and integer status returns.
+- 2026-08-30 · issue-511 · new-page · adr/skill-overlap-ratchet-005.md · retired the skill-overlap ratchet: it could never block (no `required_status_checks` on `main`, no merge queue), calibration and baseline never left `draft`, and the Rust analyzer cost a build plus model fetch per PR for an advisory report with no consumer.
+- 2026-08-30 · issue-511 · merged · adr/skill-overlap-ratchet-001.md, adr/skill-overlap-ratchet-002.md, adr/skill-overlap-ratchet-003.md, adr/skill-overlap-ratchet-004.md, tooling.md · marked the four ratchet ADRs superseded and dropped the deleted `skill-overlap.yml` workflow and the `cargo`-dependent `test-skill-overlap` gate leg; `just check` is now Python, JS, and bash only.
