+++ justfile
@@ -28,11 +28,6 @@ test:
     node --test 'tests/js/**/*.test.mjs'
     bats tests/bash/test_install.bats
     uv run --no-project --with-requirements requirements/runtime.txt --with pip==26.2.1 --with pyyaml==6.0.2 bats tests/fanout/bash/test_pr_plan_to_branches.bats
-    just test-skill-overlap
-
-# Run model-free overlap analyzer tests (never fetches model artifacts)
-test-skill-overlap:
-    cargo test --manifest-path tools/skill-overlap/Cargo.toml
 
 # Build one self-contained Shiv .pyz archive per Python skill
 bundle:
