--- tools/skill-overlap/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name = "skill-overlap"
-version = "0.1.0"
-edition = "2021"
-publish = false
-
-[dependencies]
-clap = { version = "4.5.41", features = ["derive"] }
-regex = "1.11.1"
-serde = { version = "1.0.219", features = ["derive"] }
-serde_json = "1.0.143"
-serde_yaml = "0.9.34"
-sha2 = "0.10.9"
-unicode-normalization = "0.1.24"
-pulldown-cmark = { version = "0.13.4", default-features = false }
-fastembed = { version = "=5.17.3", optional = true }
-reqwest = { version = "0.12.28", default-features = false, features = ["blocking", "rustls-tls"], optional = true }
-tokenizers = { version = "=0.22.2", optional = true }
-
-[features]
-model = ["dep:fastembed", "dep:reqwest", "dep:tokenizers"]
-
-[dev-dependencies]
-tempfile = "3"
