--- tools/skill-overlap/model.lock.json
@@ -1,20 +1,0 @@
-{
-  "format": 1,
-  "model": "snowflake/snowflake-arctic-embed-s",
-  "revision": "b637eda6144b122ccef9318e9c8dd1483399ce87",
-  "fastembed": "5.17.3",
-  "dimensions": 384,
-  "execution_provider": "onnxruntime-cpu",
-  "threads": 1,
-  "batch_size": 32,
-  "passage_prefix": "",
-  "pooling": "cls",
-  "normalization": "l2",
-  "artifacts": [
-    {"path": "onnx/model.onnx", "sha256": "579c1f1778a0993eb0d2a1403340ffb491c769247fb46acc4f5cf8ac5b89c1e1"},
-    {"path": "config.json", "sha256": "4e519aa92ec40943356032afe458c8829d70c5766b109e4a57490b82f72dcfb7"},
-    {"path": "tokenizer.json", "sha256": "91f1def9b9391fdabe028cd3f3fcc4efd34e5d1f08c3bf2de513ebb5911a1854"},
-    {"path": "tokenizer_config.json", "sha256": "de2ec2766c807b88a6442718d304a4217c2027aa93dd4060e9677bf92363abb1"},
-    {"path": "special_tokens_map.json", "sha256": "5d5b662e421ea9fac075174bb0688ee0d9431699900b90662acd44b2a350503a"}
-  ]
-}
